+++ kuber-server/src/main/java/com/kuber/server/controller/CustomErrorController.java
@@ -45,7 +45,7 @@ import java.util.UUID;
  * - Request tracking with unique IDs
  * 
  * @author Ashutosh Sinha
- * @version 1.2.2
+ * @version 1.2.4
  */
 @Slf4j
 @Controller